componentize featuredListings.js into HorizontalScrollMenu.js
